+++ addon/fields/records.js
@@ -1,5 +1,6 @@
-import { camelize, pluralize } from 'ember-cli-mirage/utils/inflector';
+import { filterRecords } from '../filter/records';
 import { getIsEdge } from '../relay/edges';
+import { getTableName } from '../db';
 import { isFunction } from '../utils';
 
 const getEdgesNodeField = (fieldName, field) =>
@@ -12,21 +13,33 @@ const getMappedFieldName = (fieldName, parent, typeName, fieldsMap = {}) =>
   parent && parent.field.type.name in fieldsMap &&
     fieldsMap[parent.field.type.name][fieldName];
 
-const getTableName = (fieldName, parent, typeName, fieldsMap) =>
-  getMappedFieldName(fieldName, parent, typeName, fieldsMap) ||
-    pluralize(camelize(typeName));
+const getFieldTableName = (fieldName, parent, typeName, fieldsMap) =>
+  getMappedFieldName(fieldName, parent, typeName, fieldsMap) || getTableName(typeName);
 
-export function getAllRecordsByType(fieldName, field, db, options) {
+export function getRecordsByField(fieldName, field, db, options) {
   let { fieldsMap = {} } = options || {};
   let edgesNodeField = getEdgesNodeField(fieldName, field);
   let typeName = (edgesNodeField || field).type.name;
-  let tableName = getTableName(fieldName, field.parent, typeName, fieldsMap);
+  let tableName = getFieldTableName(fieldName, field.parent, typeName, fieldsMap);
   let table = db[tableName] || [{}];
 
   return table.slice(0);
 }
 
-export function getRecordsByMappedFieldFn(records, field, fieldName, db, options = {}) {
+export function getRecordsInField(records, { db, field, fieldName, options, vars }) {
+  if (field.relayNode) {
+    records = [field.relayNode];
+  } else if (field.relayPageInfo) {
+    records = [field.relayPageInfo];
+  } else {
+    records = getRecordsByField(fieldName, field, db, options);
+    records = filterRecords(records, field, fieldName, vars, options);
+  }
+
+  return records;
+}
+
+export function getRecordsByMappedFieldFn(records, { db, field, fieldName, options = {} }) {
   let { fieldsMap = {} } = options;
   let fieldsMapForType = getFieldsMapForType(field.parent, fieldsMap);
   let resolvedFieldName = fieldsMapForType && fieldsMapForType[fieldName];
@@ -38,7 +51,7 @@ export function getRecordsByMappedFieldFn(records, field, fieldName, db, options
         ? field.parent.field.parent.record
         : field.parent.record
 
-    return resolvedFieldName(records, db, parent);
+    records = resolvedFieldName(records, db, parent);
   }
 
   return records;
