+++ addon/filter/sort.js
@@ -1,0 +1,3 @@
+const sortFilters = ({ hasFnValue }) => hasFnValue ? 1 : -1;
+
+export default sortFilters;
