+++ addon/relay/edges.js
@@ -1,30 +1,20 @@
-import { applyRelayFilters } from './filters';
-import { createPageInfo } from './page-info';
 import { fieldHasConnectionType } from './connection';
 
 export const getIsEdge = (fieldName, field) =>
   fieldName === 'edges' && fieldHasConnectionType(field.parent.field.type.name);
 
-const getRecordToEdgeMapper = (typeName) => (record) => ({
-  cursor: btoa(`${typeName}:${record.id}`),
-  node: record
-});
+const getRecordToEdgeMapper = (typeName) =>
+  (record) => ({
+    cursor: btoa(`${typeName}:${record.id}`),
+    node: record
+  });
 
-export function getEdgesAndPageInfo(records, filters = [], typeName) {
-  let edges = [];
-  let hasRecords = !!records.length;
-  let firstRecordId = hasRecords && records[0].id;
-  let lastRecordId = hasRecords && records[records.length - 1].id;
+export function createRelayEdges(records, { field }) {
+  let { fields, isRelayEdges } = field;
 
-  if (hasRecords) {
-    edges = applyRelayFilters(records, filters)
-      .map(getRecordToEdgeMapper(typeName));
+  if (isRelayEdges) {
+    records = records.map(getRecordToEdgeMapper(fields.node.type.name));
   }
 
-  let pageInfo = createPageInfo(edges, firstRecordId, lastRecordId, typeName);
-
-  return [edges, pageInfo];
+  return records;
 }
-
-export const createRelayEdges = (records, typeName) =>
-  records.map(getRecordToEdgeMapper(typeName));
