+++ addon/utils.js
@@ -1,5 +1,3 @@
-const pipeReducer = (f, g) => (...args) => g(f(...args));
-
 export function contextPush(context, k, v) {
   context[k].push(v);
 
@@ -12,14 +10,6 @@ export function contextSet(context, k, v) {
   return context;
 }
 
-export function cutKey(obj, k) {
-  let v = obj[k];
-
-  obj[k] = null;
-
-  return v;
-}
-
 export const ensureList = (item) => !item
   ? []
   : item instanceof Array
@@ -30,7 +20,8 @@ export const isFunction = (obj) => obj != null && typeof obj === 'function';
 
 export const partial = (fn, ...args1) => (...args2) => fn(...args1, ...args2);
 
-export const pipe = (...fns) => fns.reduce(pipeReducer);
+export const pipeWithMeta = (...fns) =>
+  fns.reduce((f, g) => (a, meta) => g(f(a, meta), meta));
 
 export const reduceKeys = (obj, reducerFn, defaultValue) =>
   Object.keys(obj).reduce(reducerFn, defaultValue);
